GitWebhook: GitLab: Detect when branches are deleted
GitLab will send a "push" event with 0 commits on branch deletion, but
this can also mean that a new branch, which is identical to master, was
created.

This works by checking if the field "checkout_sha" is null. GitLab
doesn't provide any better way to achieve this and does not document
this behavior, so this might break in the future.