Rename IRCColorCodes and IRCColorsHex
While they are still based on the IRC colors (and will always remain
that way), they should serve as a generic set of colors used for any
backend.
For Matrix, they shall in the future be sent as named colors and not
RGB/Hex values, so that they're not tied to a color scheme.