GitWebhook: Support namespaces for report channel configuration
If there is no "/" in the config key, only the repo name is matched.
If a "/" is present, the namespace is taken into account. For GitLab,
this has to be the full namespace containing all subgroups.
A "*" can be used as default value for both namespace and repo name.
THERE IS NO GLOBBING!